обновление окна инвентаря
- удален прогресс-бар алкоголя
- удален прогресс-бар выносливости
+ добавлено визуальное отражение блокировки слотов ПНВ и противогаза.
~ доработан пояс, зависимый от костюма, при снятии брони артефакты
сбрасываются в рюкзак, заблокированные ячейки блокируются статиком.
~ поправлены навзания некоторых слотов на более вменяемые
~ поправлены названия некоторых листов на более вменяемые
~ вырезано окно сна из инвентаря
~ слот ITEMS_SLOT теперь не имеет визуального отражения в рюкзаке
~ удалена криво перенесенная из ЗП функция наполнения слотов-листов.
Сделано как и было В ТЧ
+ добавленая возможность задавать отступы между клетками в
DragDropListEx
~ продолжена работа над фонарем. Удалена бесполезная функция апдейта
светя скопипастеная из обычного (делать надо иначе)
+ добавлена поддержка объемного света у ручного фонаря.

+++ trunk/xray/xr_3da/xrGame/ui/UIMainIngameWnd.cpp
@@ -600,7 +600,7 @@ void CUIMainIngameWnd::Update()
 		}
 	}
 // Weapon broken icon
-	u16 slot = m_pActor->inventory().GetActiveSlot();
+	u32 slot = m_pActor->inventory().GetActiveSlot();
 	m_ind_weapon_broken->Show(false);
 	m_ind_weapon_broken->ResetClrAnimation();
 	if(slot == PISTOL_SLOT || slot == RIFLE_SLOT)
